Performance optimization for methods with multiple returns
The analyzer for AV1250 now caches the evaluated state of variables, so they can be reused between multiple return statements within the same method.
This speeds up analysis of methods with multiple returns.

For example:

```csharp
IEnumerable<int> M(IList<int> source, bool flag)
{
    var result1 = Enumerable.Empty<int>();
    result1 = source.Select(x => x);

    var result2 = result1;

    var result3 = result2;

    if (flag)
    {
        return result2;
    }
    else
    {
        return result3;
    }
}
```

In this example, analysis for `result3` reuses the cached evaluation state of `result2`, so it does not need to re-evaluate the `result1` declaration and assignment.